Refactored how ListSchedule.jsx renders table
