+++ client/app/hearings/components/dailyDocket/DailyDocket.jsx
@@ -260,8 +260,8 @@ export default class DailyDocket extends React.Component {
               <React.Fragment>
                 `VLJ: ${dailyDocket.judgeFirstName} ${dailyDocket.judgeLastName}`
                 <br />
-               </React.Fragment>
-             )}
+              </React.Fragment>
+            )}
             Coordinator: {dailyDocket.bvaPoc} <br />
             Hearing type: {dailyDocket.readableRequestType} <br />
             Regional office: {dailyDocket.regionalOffice}
